bot/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
discord.py==1.7.2
idna==2.10
multidict==5.1.0
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
python-dotenv==0.18.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
